#19 feat: service -> repositoryService로 변경
test 코드 삭제
