--- src/main/java/com/example/recommendationservice/domain/pharmacy/service/PharmacyService.java
@@ -1,60 +1,0 @@
-package com.example.recommendationservice.domain.pharmacy.service;
-
-import com.example.recommendationservice.domain.pharmacy.entity.Pharmacy;
-import com.example.recommendationservice.domain.pharmacy.repository.PharmacyRepository;
-import java.util.List;
-import java.util.Objects;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.transaction.support.TransactionSynchronizationManager;
-
-@Slf4j
-@Service
-@RequiredArgsConstructor
-public class PharmacyService {
-
-    private final PharmacyRepository pharmacyRepository;
-
-    // self invocation sovle
-    @Transactional
-    public void bar(List<Pharmacy> pharmacyList) {
-        log.info("bar CurrentTransactionName: " + TransactionSynchronizationManager.getCurrentTransactionName());
-        foo(pharmacyList);
-    }
-
-    // self invocation sovle
-    public void foo(List<Pharmacy> pharmacyList) {
-        log.info("foo CurrentTransactionName: " + TransactionSynchronizationManager.getCurrentTransactionName());
-        pharmacyList.forEach(pharmacy -> {
-            pharmacyRepository.save(pharmacy);
-            throw new RuntimeException("throw error");
-        });
-    }
-
-
-    @Transactional
-    public void updateAddress(Long id, String address) {
-        Pharmacy entity = pharmacyRepository.findById(id).orElse(null);
-
-        if (Objects.isNull(entity)) {
-            log.error("[PharmacyService updateAddress] not found id : {}", id);
-            return;
-        }
-
-        entity.changePharmacyAddress(address);
-    }
-
-    // for test
-    public void updateAddressWithoutTransaction(Long id, String address) {
-        Pharmacy entity = pharmacyRepository.findById(id).orElse(null);
-
-        if (Objects.isNull(entity)) {
-            log.error("[PharmacyService updateAddress] not found id : {}", id);
-            return;
-        }
-
-        entity.changePharmacyAddress(address);
-    }
-}
